refactor(preproc): split source model modules
